+++ pothole-manager-api/src/main/java/pothole_solution/manager/global/config/EnumFormatter.java
@@ -4,12 +4,12 @@ import org.jetbrains.annotations.NotNull;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import pothole_solution.manager.global.converter.ReportPeriodRequestConverter;
+import pothole_solution.manager.global.converter.ReportPeriodEnumConverter;
 
 @Configuration
 public class EnumFormatter implements WebMvcConfigurer {
     @Override
     public void addFormatters(@NotNull FormatterRegistry registry) {
-        registry.addConverter(new ReportPeriodRequestConverter());
+        registry.addConverter(new ReportPeriodEnumConverter());
     }
 }
