+++ pothole-manager-api/src/main/java/pothole_solution/manager/global/converter/ReportPeriodEnumConverter.java
@@ -1,9 +1,10 @@
-package pothole_solution.manager.service;
+package pothole_solution.manager.global.converter;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.core.convert.converter.Converter;
+import pothole_solution.manager.report.entity.ReportPeriod;
 
-public class ReportPeriodRequestConverter implements Converter<String, ReportPeriod> {
+public class ReportPeriodEnumConverter implements Converter<String, ReportPeriod> {
     @Override
     public ReportPeriod convert(@NotNull String period) {
         return ReportPeriod.enumOf(period);
